+++ examples/sampler_benchmarking/run_all.py
@@ -26,7 +26,7 @@ ALL_PROBLEMS = [
     "TCR", "FcERI_gamma", "MEK_Isoforms",
     "EGFR_d10", "EGFR_d37",
 ]
-ALL_SAMPLERS = ["am", "dream", "p_dream", "s_cream"]
+ALL_SAMPLERS = ["am", "dream", "p_dream"]
 
 
 def build_local_command(conf_file, uses_bng=False):