+++ pybnf/config.py
@@ -140,7 +140,7 @@ class Configuration(object):
             d = self.check_unused_keys_model_checking(d)
         elif verbosity >= 1:
             self.check_unused_keys(d)
-        if d['fit_type'] in ('mh', 'pt', 'sa', 'dream', 'p_dream', 's_cream', 'am'):
+        if d['fit_type'] in ('mh', 'pt', 'sa', 'dream', 'p_dream', 'am'):
             self.postprocess_mcmc_keys(d)
         self.config = self.default_config()
         for k, v in d.items():
@@ -228,8 +228,6 @@ class Configuration(object):
             'delta': 1,
             'outlier_method': 'iqr',
             'rhat_threshold': 0.0,
-            'refset_size': None,
-            'refset_quality_fraction': 0.5,
             'precondition_adapt': None
         }
         return default
@@ -251,7 +249,6 @@ class Configuration(object):
                                 'crossover_number', 'zeta', 'lambda', 'gamma_prob', 'adaptive', 'time_length',
                                 'archive_size', 'archive_thin_rate', 'snooker_prob',
                                 'delta', 'outlier_method', 'rhat_threshold',
-                                'refset_size', 'refset_quality_fraction',
                                 'precondition_adapt'},
                         'sim': {'simplex_step', 'simplex_log_step', 'simplex_reflection', 'simplex_expansion',
                                 'simplex_contraction', 'simplex_shrink', 'simplex_max_iterations',
@@ -259,7 +256,7 @@ class Configuration(object):
                         }
         ignored_params = set()
         thisalg = conf_dict['fit_type']
-        if thisalg in ('pt', 'sa', 'dream', 'p_dream', 's_cream', 'am'):
+        if thisalg in ('pt', 'sa', 'dream', 'p_dream', 'am'):
             thisalg = 'mh'
         for alg in alg_specific:
             if (thisalg != alg